requirements.txt
streamlit
openai
python-dotenv
PyPDF2
requests
vosk
pyaudio
wave
streamlit-webrtc